auth/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib
python-jose
pydantic
streamlit
requests
langchain
langchain-openai
pytesseract
python-docx
PyMuPDF
Pillow

